Fixing issue where out_of_bounds would allow index out of range errors
Out of bounds was allowing access of indices equal to the number of
map rows and columns.

Because c++ has zero-based numbering, an attempt to access element
map_rows of the map 2D array would result in a segmentation fault.

This behaviour was masked, as it would only occur if an Actor moved all
the way to the right (and tests were done with Actors moving to the left)
or if an Actor happened to be randomly initialized at the edge.